requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
fastapi-mail==1.4.1
python-dotenv==1.0.0
aiosmtplib==2.0.2
blinker==1.9.0

# Database
psycopg2-binary==2.9.9
alembic==1.12.1

# HTTP clients
httpx==0.25.1
requests==2.31.0

# Templates and static files
jinja2==3.1.2
aiofiles==23.2.1

# Server
gunicorn==21.2.0
httptools==0.6.1

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Scheduling
APScheduler==3.10.1

# Validation
email-validator==2.1.0.post1
